stack-z-slices/src/requirements.txt
python_bioformats==1.5.2
javabridge==1.0.18
bfio==1.0.1
